% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Creation_v0.6.R
\name{ArrayToAsc}
\alias{ArrayToAsc}
\title{ArrayToAsc converts an array to asc format modifying and stitching a header on.}
\usage{
ArrayToAsc(inArray, filename, overwrite = FALSE, runno = NULL)
}
\arguments{
\item{inArray}{array - an input array that will provide cell values for the asc file.}

\item{filename}{character - a string giving the filename and path.}

\item{overwrite}{boolean - by default set as NULL. If this is set to true, then the file
can overwrite another ascii file with same name.}

\item{runno}{boolean - as part of the process a tempory file is created. If multiple
ascii files are being created simultanously then an identifier may stop a conflict of temp
files.}
}
\value{
character - currently returns the filepath
}
\description{
This adds on a simple header to an array to create an asc spatial dataset and saves
this to the hard drive.
}
\examples{

TBA...

}
