% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Creation_v0.6.R
\name{CreateWorld}
\alias{CreateWorld}
\title{CreateWorld sets up an initial matrix with the focal patches in}
\usage{
CreateWorld(SeedPoints, PercentageCover, Extent, SeedDistribution = NULL,
  PatchDistribution = NULL, SecondPercentageCover = 0,
  normalSD = 0.01, paretoVal = 1)
}
\arguments{
\item{SeedPoints}{integer - number of patches}

\item{PercentageCover}{float between 0.0 & 1.0 giving the proportion of array that should be patches.}

\item{Extent}{integer - size of square array, defaults to 250 (large arrays take longer to create.)}

\item{SeedDistribution}{character - how seed points are distributed to grow the patches. Options are "Random" or "Poisson", "Clustered" or "Thomas", or "Regular", "Strauss" or"Dispersed"
and defaults to random.}

\item{PatchDistribution}{character -a distribution from which to pick the size of patches from. Opetions are
"Gaussian" or"Normal", "Pareto", or the default "Uniform". Uniform gives patches of almost any size and the maximum variation.}

\item{SecondPercentageCover}{float between 0.0 & 1.0 giving the proportion of array that should be patches.If > 0 creates an additional array based on the seeds of the first with a different proportion.}

\item{normalSD}{default = 0.01,}

\item{paretoVal}{defulat = 1}
}
\value{
list of arrays - the list contains two arrays, the updated GrowWorld and WeightWorld.
}
\description{
creates a landscape with zeros in and then adds patches
}
\examples{

TBA...
}
