% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Creation_v0.6.R
\name{ExtractWindow}
\alias{ExtractWindow}
\title{ExtractWindow extracts a defined region of an array.}
\usage{
ExtractWindow(InMatrix, y, x, ySize = 3, xSize = 3)
}
\arguments{
\item{InMatrix}{array - an input array from which a subset will be taken.}

\item{y}{integer - central y location for subset.}

\item{x}{integer - central x location for subset.}

\item{ySize}{integer - height of window}

\item{xSize}{integer - width of window}
}
\value{
array - an array smaller than the input array. Of the 3 by 3 by default or larger if
specified.
}
\description{
This extracts a number of cells from an array. Typically used in a moving window for example.
}
\examples{

TBA...

}
