% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Creation_v0.6.R
\name{FillTheMatix_specific}
\alias{FillTheMatix_specific}
\title{FillTheMatix_specific adds habitats of specified type to the space between
habitat patches.}
\usage{
FillTheMatix_specific(Patches, InterestLCV, PossMatrixHabs = seq(2, 11),
  PossibleNumberMatrixsPoints = 1:200, EqualSize = FALSE,
  ChosenPatchDistribution = "Random", SeedDistribution = "Uniform")
}
\arguments{
\item{Patches}{array -}

\item{InterestLCV}{array -}

\item{PossibleNumberMatrixsPoints}{list of integers -}

\item{EqualSize}{boolean -}

\item{ChosenPatchDistribution}{character - how seed points are distributed to grow the patches. Options are "Random" or "Poisson", "Clustered" or "Thomas", or "Regular", "Strauss" or"Dispersed"
and defaults to random.}

\item{SeedDistribution}{character -a distribution from which to pick the size of patches from. Opetions are
"Gaussian" or"Normal", "Pareto", or the default "Uniform". Uniform gives patches of almost any size and the maximum variation.}

\item{MatrixHabs}{integer - default 10 habitats}
}
\value{
list of arrays - the list contains two arrays, the "Patches" and "InterestLCV" (land cover).
}
\description{
This function adds habitats into the intervening space between a focal habitat patch.
It differs from \link[LcvGen]{FillTheMatix} by allowing non sequential habitats. For example,
habitat 2, 5, 7, 8, 9, 100.
}
\examples{

TBA...

}
