% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Creation_v0.6.R
\name{GenerateLandcover}
\alias{GenerateLandcover}
\title{GenerateLandcover creates a ascii land cover dataset with a corresponding patch number data set.}
\usage{
GenerateLandcover(NumberSeeds, LCVPath, PatchPath, runno = NULL,
  MatrixHabs = 10, GenExtent = 250, FullExtent = 1000,
  BufferFullExtent = 0, cover = 0.1, DistributionPatches = "Random",
  DistributionSeeds = "Uniform", NumberEachMatrixHabs = 1:200,
  MatrixHabsEqualSize = FALSE, MatrixPatchDistribution = "Random",
  MarixSeedDistribution = "Uniform")
}
\arguments{
\item{NumberSeeds}{integer -}

\item{LCVPath}{character - string of the output path of the land cover ascii}

\item{PatchPath}{character - string of the output path of the patch ascii}

\item{runno}{integer - defaults to NULL}

\item{MatrixHabs}{integer -number of habitats in the matrix}

\item{GenExtent}{integer - default 250, this is the extent that the land cover is generated at.}

\item{SeedDistribution}{character -a distribution from which to pick the size of patches from. Opetions are
"Gaussian" or"Normal", "Pareto", or the default "Uniform". Uniform gives patches of almost any size and the maximum variation.}
}
\value{
list of arrays - the list contains two arrays, the "Patches" and "InterestLCV" (land cover).
}
\description{
This function uses other functions to create a land cover with the focal habitat haveing a number
of patches.
}
\examples{

TBA...

}
