% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Creation_v0.6.R
\name{GrowBuffer}
\alias{GrowBuffer}
\title{GrowBuffer adds a cell of habitat when growing habitat patches.}
\usage{
GrowBuffer(yLocal, xLocal, seed, Extent, World, GrowWorld, WeightWorld)
}
\arguments{
\item{yLocal}{array - an input array from which a subset will be taken.}

\item{xLocal}{integer - central y location for subset.}

\item{seed}{integer - central x location for subset.}

\item{Extent}{integer - height of window}

\item{World}{integer - width of window}

\item{GrowWorld}{array that defines where the patches of habitat are.}

\item{WeightWorld}{array of a weighted buffer around the patches used to grow patches.}
}
\value{
list of arrays - the list contains two arrays, the updated GrowWorld and WeightWorld.
}
\description{
This takes two input arrays, one with the filled patches and the other with the buffered patches
and converts one of the buffer cells into a cell of habitat and expands the buffer.I think
it weights the buffer so that it often fills holes first.
}
\examples{

TBA...

}
