% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Creation_v0.6.R
\docType{data}
\name{header}
\alias{header}
\title{header is the header for a asc file. This is a bit of a bodge, but can be added to
an array to give and ascii file.}
\format{
Six character objects defining an ascii
}
\usage{
The data is automatically used and has no real independent use.
}
\description{
header is the header for a asc file. This is a bit of a bodge, but can be added to
an array to give and ascii file.
}
\keyword{datasets}
