% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Creation_v0.6.R
\docType{package}
\name{LcvGen-package}
\alias{LcvGen}
\alias{LcvGen-package}
\title{Package for creating simulated ascii land covers}
\description{
This package allows for presicie control of generation of focal habitats
as well as filling in the matrix habitats inbetween the focal patches.
}
\details{
The main function i is \link[PhiCor]{CorIndex.all.plusP}.
    This will calculate all phi corelation of association values and
    purmutate to give p-values.

    The other Phi coefficent functions are useful in testing or calculating only part
    of the full outputs from \link[PhiCor]{CorIndex.all.plusP}:
    \itemize{
        \item \link[PhiCor]{CorIndex} calculates the inputs that don't change between habitats.
        \item \link[PhiCor]{CorIndex.TargetVar} calculates the additional input specific to a habitat.
        \item \link[PhiCor]{CorIndex.groupEqual} calculates the group-equalised phi for a habitat.
        \item \link[PhiCor]{CorIndex.nonEqual} calculates the non group-equalised phi for a habitat.
        \item \link[PhiCor]{CorIndex.xPerm} returns the null phi distribution for a habitat.
        \item \link[PhiCor]{CorIndex.all} returns the non and group-equalised phi values for all habitats
            without p-values and is therefore quicker.
    }


    The functions \link[PhiCor]{lookup} and \link[PhiCor]{FilePathInTree}
    might be used elsewhere.
}
\references{



}
\author{
Jordan Chetcuti
}
\keyword{internal}
