% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Creation_v0.6.R
\name{GrowBuffer}
\alias{GrowBuffer}
\title{GrowBuffer adds a cell of habitat when growing habitat patches.}
\usage{
GrowBuffer(yLocal, xLocal, seed, Extent, World, GrowWorld, WeightWorld)
}
\arguments{
\item{yLocal}{array - an input array from which a subset will be taken.}

\item{xLocal}{integer - central y location for subset.}

\item{seed}{integer - central x location for subset.}

\item{Extent}{integer - height of window}

\item{World}{integer - width of window}

\item{GrowWorld}{array}

\item{WeightWorld}{array}
}
\value{
list of arrays - the list contains two arrays, the updated GrowWorld and WeightWorld.
}
\description{
I wrote this and I've lost track of exactly what it does...
}
\examples{

TBA...

}
