% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Creation_v0.6.R
\docType{package}
\name{LcvGen-package}
\alias{LcvGen}
\alias{LcvGen-package}
\title{Package for creating simulated ascii land covers}
\description{
This package allows for presicie control of generation of focal habitats
as well as filling in the matrix habitats inbetween the focal patches.
}
\details{
The main function i are \link[LcvGen]{GenerateLandcover} and pottentially
    \link[LcvGen]{CreateWorld} and \link[LcvGen]{FillTheMatix}.
    This will generate a land cover and patches ascii datasets.

    Many of the other LCVGen functions are used in \link[LcvGen]{GenerateLandcover}.
    Some of them may be useful alone however:
    \itemize{
        \item \link[LcvGen]{CreateWorld} -
        \item \link[LcvGen]{CreatSeeds} -
        \item \link[LcvGen]{AddAllSeedPoints} -
        \item \link[LcvGen]{ExtractWindow} -
        \item \link[LcvGen]{GrowBuffer} -
        \item \link[LcvGen]{CreateAdditionalWorld} -
        \item \link[LcvGen]{AddMoreSeedPoints}
        \item \link[LcvGen]{GrowAllBuffer}
        \item \link[LcvGen]{FillAllZeroPatches}
        \item \link[LcvGen]{ScaleUpArray}
        \item \link[LcvGen]{AddBuffer}
        \item \link[LcvGen]{ArrayToAsc}
    }


    The functions ? and ?
    might be used elsewhere.
}
\references{



}
\author{
Jordan Chetcuti
}
\keyword{internal}
